airflow/requirements.txt
psycopg2-binary==2.9.10
jinja2==3.1.6